modified analytics and dashboards app, now they serve their right purpose and are more separated and functional
