+++ amadeus/templates/base.html
@@ -251,7 +251,7 @@
                             </a>
                         </li>
                         <li class="item" data-toggle="tooltip" data-placement="right" title="{% trans "Analytics" %}">
-                            <a href="{% url 'analytics:view_general' %}"><i class="fa fa-bar-chart" aria-hid{% get_current_language as LANGUAGE_CODE %}den="true"></i></a>
+                            <a href="{% url 'dashboards:view_general' %}"><i class="fa fa-bar-chart" aria-hid{% get_current_language as LANGUAGE_CODE %}den="true"></i></a>
                         </li>
                     </ul>
                 {% endblock %}
@@ -306,7 +306,7 @@
                             </a>
                         </li>
                         <li class="item" data-toggle="tooltip" data-placement="top" title="{% trans "Analytics" %}">
-                            <a href="{% url 'analytics:view_general' %}"><i class="fa fa-bar-chart" aria-hid{% get_current_language as LANGUAGE_CODE %}den="true"></i></a>
+                            <a href="{% url 'dashboards:view_general' %}"><i class="fa fa-bar-chart" aria-hid{% get_current_language as LANGUAGE_CODE %}den="true"></i></a>
                         </li>
                     </ul>
                 </div>
