+++ dashboards/static/dashboards/js/behavior.js
