+++ dashboards/templates/dashboards/general.html
@@ -9,13 +9,13 @@
 {% block javascript %}
 
     <script type="text/javascript" src="{% static "analytics/js/charts.js" %}"></script>
-    <script type="text/javascript" src=" {% static "analytics/js/behavior.js" %} "></script>
+    <script type="text/javascript" src=" {% static "dashboards/js/behavior.js" %} "></script>
 {% endblock javascript %}
 
 {% block breadcrumbs %}
     {{ block.super }}
     
-    {% trans 'Analytics General' as general %}
+    {% trans 'Dashboards General' as general %}
 
     {% breadcrumb general 'dashboard:view_general' %}
 {% endblock %}
@@ -29,9 +29,9 @@
                     <li class="active">
                         {% trans "Platform Report" %}
                     </li>
-                    <li>
+                   <a href=""><li>
                         {% trans "Category Report" %}
-                    </li>
+                    </li></a> 
                     <li>
                         {% trans "Subject Report" %}
                     </li>
